refactor : 임시 데이터 모두 get함수로 교체했습니다 컴포넌트 분리했습니다
